+++ bindgen-tests/tests/headers/wrap_unsafe_ops_dynamic_loading_simple.h
@@ -1,0 +1,5 @@
+// bindgen-flags: --dynamic-loading TestLib --wrap-unsafe-ops
+
+int foo(int x, int y);
+int bar(void *x);
+int baz();