+++ bindgen/codegen/dyngen.rs
@@ -1,4 +1,5 @@
 use crate::codegen;
+use crate::ir::context::BindgenContext;
 use crate::ir::function::ClangAbi;
 use proc_macro2::Ident;
 
@@ -72,11 +73,21 @@ impl DynamicItems {
         Self::default()
     }
 
-    pub fn get_tokens(&self, lib_ident: Ident) -> proc_macro2::TokenStream {
+    pub fn get_tokens(
+        &self,
+        lib_ident: Ident,
+        ctx: &BindgenContext,
+    ) -> proc_macro2::TokenStream {
         let struct_members = &self.struct_members;
         let constructor_inits = &self.constructor_inits;
         let init_fields = &self.init_fields;
         let struct_implementation = &self.struct_implementation;
+
+        let from_library = if ctx.options().wrap_unsafe_ops {
+            quote!(unsafe { Self::from_library(library) })
+        } else {
+            quote!(Self::from_library(library))
+        };
 
         quote! {
             extern crate libloading;
@@ -92,7 +103,7 @@ impl DynamicItems {
                 ) -> Result<Self, ::libloading::Error>
                 where P: AsRef<::std::ffi::OsStr> {
                     let library = ::libloading::Library::new(path)?;
-                    Self::from_library(library)
+                    #from_library
                 }
 
                 pub unsafe fn from_library<L>(
@@ -124,6 +135,7 @@ impl DynamicItems {
         ret: proc_macro2::TokenStream,
         ret_ty: proc_macro2::TokenStream,
         attributes: Vec<proc_macro2::TokenStream>,
+        ctx: &BindgenContext,
     ) {
         if !is_variadic {
             assert_eq!(args.len(), args_identifiers.len());
@@ -164,13 +176,19 @@ impl DynamicItems {
 
         // N.B: Unwrap the signature upon construction if it is required to be resolved.
         let ident_str = codegen::helpers::ast_ty::cstr_expr(ident.to_string());
+        let library_get = if ctx.options().wrap_unsafe_ops {
+            quote!(unsafe { __library.get(#ident_str) })
+        } else {
+            quote!(__library.get(#ident_str))
+        };
+
         self.constructor_inits.push(if is_required {
             quote! {
-                let #ident = __library.get(#ident_str).map(|sym| *sym)?;
+                let #ident = #library_get.map(|sym| *sym)?;
             }
         } else {
             quote! {
-                let #ident = __library.get(#ident_str).map(|sym| *sym);
+                let #ident = #library_get.map(|sym| *sym);
             }
         });